data_lake/pasx-savvy-sync/container/requirements.txt
ConfigArgParse==1.5.3
coloredlogs==15.0.1
psycopg2-binary==2.9.3
requests==2.27.1
